requirements.txt
black==22.6.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
fonttools==4.33.3
importlib-metadata==4.12.0
kiwisolver==1.4.3
matplotlib==3.5.2
mypy-extensions==0.4.3
numpy==1.21.6
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
zipp==3.8.0
